# backend-filemanagement/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "List all plans",
                "description": "Returns a list of all plans",
                "operationId": "indexPlans",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "api/plans/store": {
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "description": "Creates a new plan",
                "operationId": "storePlan",
                "requestBody": {
                    "description": "Plan data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully"
                    }
                }
            }
        },
        "api/plans/show/{id}": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Show plan by ID",
                "description": "Returns a plan by its ID",
                "operationId": "showPlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plan to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "api/plans/update/{id}": {
            "put": {
                "tags": [
                    "Plans"
                ],
                "summary": "Update an existing plan",
                "description": "Update an existing plan",
                "operationId": "updatePlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plan to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Plan data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Plan updated successfully"
                    }
                }
            }
        },
        "api/secret-keys": {
            "get": {
                "tags": [
                    "Secret Keys"
                ],
                "summary": "List all secret keys",
                "description": "Returns a list of all secret keys",
                "operationId": "indexSecretKeys",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}